+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+### Added
+- `casing` utility to `forr!`
+
 ## [0.2.2] - 2023-10-28
 ### Added
 - support for `#` instead of `$` inside of macros, when they are initialized with `#:`/`#*`